Stats: Implement exchangeability blocks
As per the description in Winkler et al., 2014.
Inputs can be identified as having exchangeability that is either restricted to occur -within- individual blocks of observations, or for blocks containing equal numbers of observations where the -whole- blocks are exchanged.
In the case of whole-block exchangeability, sign-flipping of independent and symmetric errors can also be restricted to apply to whole blocks rather than individual observations.

+++ core/math/stats/typedefs.h
@@ -32,7 +32,7 @@ namespace MR
       using value_type = MR::default_type;
       using matrix_type = Eigen::Matrix<value_type, Eigen::Dynamic, Eigen::Dynamic>;
       using vector_type = Eigen::Array<value_type, Eigen::Dynamic, 1>;
-      using index_array_type = Eigen::Array<int, Eigen::Dynamic, 1>;
+      using index_array_type = Eigen::Array<size_t, Eigen::Dynamic, 1>;
